introduce possibility to use pip packages that depend on libxmlsec1
dm.xmlsec.binding, which comes as a dep for python-saml,
does not compile on Heroku, using this buildpack, because
apparently /usr/include is not available at that time.

After some research, it seems that is known behavior and the reason
why workarounds for some packages (e.g.sqlite [1]) exist.

Hence it may be easiest to simply do something very similiar for
libxmlsec1, hopefully without breaking anything else.

[1] https://github.com/heroku/heroku-buildpack-python/pull/858